Add BlogBlock refactor
