+++ frontend/src/utils/data-fetching.js
@@ -1,0 +1,5 @@
+/* custom headers used for POST requests */
+export const headers = {
+  "X-CSRFToken": window.CSRF_TOKEN || "",
+  "Content-Type": "application/json",
+}
