+++ frontend/src/utils/icons.js
@@ -1,0 +1,9 @@
+export const getTypeIcon = (type) => {
+  const mapping = {
+    plant: "ri-plant-line",
+    microorganism: "ri-microscope-line",
+    ingredient: "ri-flask-line",
+    substance: "ri-test-tube-line",
+  }
+  return mapping[type] || null
+}
