+++ frontend/src/views/LandingPage/BlogBlock.vue
@@ -1,5 +1,5 @@
 <template>
-  <div v-if="blogPosts && blogPosts.length > 0">
+  <div v-if="blogPosts.results.length > 0">
     <h3>Les ressources nouvellement ajoutées</h3>
     <div class="mb-4 text-right">
       <router-link :to="{ name: 'BlogsHome' }">
@@ -9,25 +9,19 @@
     </div>
 
     <div class="grid grid-cols-12 gap-4">
-      <BlogCard class="col-span-12 sm:col-span-6 md:col-span-4" v-for="post in blogPosts" :key="post.id" :post="post" />
+      <BlogCard
+        class="col-span-12 sm:col-span-6 md:col-span-4"
+        v-for="post in blogPosts.results"
+        :key="post.id"
+        :post="post"
+      />
     </div>
   </div>
 </template>
 
 <script setup>
-import { ref, onMounted } from "vue"
-import { verifyResponse } from "@/utils"
 import BlogCard from "@/components/BlogCard"
+import { useFetch } from "@vueuse/core"
 
-const blogPosts = ref(null)
-
-onMounted(() => {
-  const url = "/api/v1/blogPosts/?limit=3&offset=0"
-  return fetch(url)
-    .then(verifyResponse)
-    .then((response) => (blogPosts.value = response.results))
-    .catch(() => {
-      window.alert("Une erreur est survenue veuillez réessayer plus tard")
-    })
-})
+const { data: blogPosts } = await useFetch("/api/v1/blogPosts/?limit=3&offset=0").json()
 </script>
