+++ frontend/src/views/ProducerForm/ProductStep.vue
@@ -148,7 +148,7 @@
 </template>
 <script setup>
 import { ref, onMounted } from "vue"
-import { verifyResponse } from "@/utils"
+import { verifyResponse } from "@/utils/custom-errors"
 
 const payload = ref({
   effects: [],
