+++ frontend/src/views/SearchResults/ReportIssue.vue
@@ -30,7 +30,8 @@ import { ref } from "vue"
 import { useVuelidate } from "@vuelidate/core"
 import { required, email, helpers } from "@vuelidate/validators"
 import { useFetch } from "@vueuse/core"
-import { headers, firstErrorMsg } from "@/utils"
+import { headers } from "@/utils/data-fetching"
+import { firstErrorMsg } from "@/utils/forms"
 import useToaster from "@/composables/use-toaster"
 
 // Props
